feat: Implement Discord message and verification services
- Added DiscordMessageService for handling wallet verification messages in Discord channels.
- Introduced DiscordVerificationService to manage user verification processes, including role assignment and nonce handling.
- Refactored DiscordService to utilize the new message and verification services, improving code organization and readability.
- Updated VerifyService to leverage DiscordVerificationService for verification logic, enhancing modularity.
